Switch backing datastructure to BTreeMap
This is to prepare for adding delete support. We can't continue to use a
Vec, since deleting a row would change the rowid for all subsequent
rows, which would mess up all the indices. The new storage mechanism is
more in line with what "real" databases use. See for example SQLite:
https://www.sqlite.org/lang_createtable.html#rowid